requirements.txt
# Core
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
xformers
trl
peft
accelerate
bitsandbytes

# Data & utilities
datasets>=2.18
scikit-learn>=1.4
pandas>=2.0
pyyaml>=6.0
torch>=2.1
